refactor: move crop endpoints to plugin
