+++ plugins/crop/README.md
@@ -1,0 +1,4 @@
+# Crop Plugin
+
+Provides perspective-crop functionality through `/detect-corners/` and `/process-image/` endpoints.
+Uploads are encrypted per-session before processing, mirroring the behavior from the main application.
